agents/crewai/requirements.txt
autogenstudio
pyautogen
crewai
crewai[tools]
python-dotenv
langchain
langchain-openai